Only search the last 3 months when looking for new runs
This is both a performance improvement (lets not search the entire
history to add 2 runs) and a short-term fix for a current outage where
Edge is perfectly aligning to an empty list on its last search page. If
we only search 90 days, we're never going to hit the last search page
anyway.